Zernike fitting : Initial commit
The zernike.py file contains the modules to fit the aperture surface
with Zernike polynomials. It also contains a class to convert the
existing CSV files to Zarr. This class will later be expanded to write
the fitted coefficients into either CSV or Zarr.

At the moment, the pipes work - i.e., the aperture gets passed into the
fitter and the fitter runs. However the results are all NaNs, likely
because of incorrectly applied NaNs